@@ -1,5 +1,5 @@
 /**
  * Copyright 2016 Richard Schoeller
- * Data Access interfaces for GEDbrowser 2.
+ * Data access objects for gedbrowser.
  */
 package org.schoellerfamily.gedbrowser.persistence;
