@@ -1,6 +1,6 @@
 /**
  * Copyright 2013-2016 Richard Schoeller
- * Data model for GEDbrowser 2.
+ * Data model for gedbrowser.
  */
 package org.schoellerfamily.gedbrowser.datamodel;
 
