@@ -1,5 +1,5 @@
 /**
  * Copyright 2016 Richard Schoeller
- * Geographic tools for GEDbrowser 2.
+ * Tests of geographic tools for gedbrowser.
  */
 package org.schoellerfamily.gedbrowser.geographic.test;
