@@ -43,73 +43,113 @@ public abstract class GedDocumentMongo<G extends GedObject>
     @Transient
     private G gedObject;
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public final String getIdString() {
         return idString;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public final void setIdString(final String idString) {
         this.idString = idString;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public final String getType() {
         return type;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public final void setType(final String type) {
         this.type = type;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public final String getString() {
         return string;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public final void setString(final String string) {
         this.string = string;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public final String getFilename() {
         return filename;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public final void setFilename(final String filename) {
         this.filename = filename;
     }
 
-
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public final String getDbName() {
         return dbName;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public final void setDbName(final String dbName) {
         this.dbName = dbName;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public final G getGedObject() {
         return gedObject;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public final void setGedObject(final G gedObject) {
         this.gedObject = gedObject;
     }
 
-
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public final List<GedDocument<? extends GedObject>> getAttributes() {
         return attributes;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public final void setAttributes(
             final List<GedDocument<? extends GedObject>> attributes) {
@@ -117,11 +157,17 @@ public abstract class GedDocumentMongo<G extends GedObject>
         this.attributes.addAll(attributes);
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public final void clearAttributes() {
         attributes.clear();
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public final void loadAttributes(final List<GedObject> gedAttributes) {
         this.attributes.clear();
@@ -137,6 +183,9 @@ public abstract class GedDocumentMongo<G extends GedObject>
         }
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
-    public abstract void loadGedObject(final GedObject ged);
+    public abstract void loadGedObject(GedObject ged);
 }
