@@ -1,5 +1,5 @@
 /**
  * Copyright 2013-2016 Richard Schoeller
- * MongoDB Data Access classes for GEDbrowser 2.
+ * Test fixture for MongoDB data access classes for gedbrowser.
  */
 package org.schoellerfamily.gedbrowser.persistence.mongo.fixture;
