@@ -1,5 +1,5 @@
 /**
  * Copyright 2013-2016 Richard Schoeller
- * MongoDB Data Access classes for GEDbrowser 2.
+ * Tests for MongoDB repository classes for gedbrowser.
  */
 package org.schoellerfamily.gedbrowser.persistence.mongo.repository.test;
