@@ -1,6 +1,6 @@
 /**
  * Copyright 2013-2016 Richard Schoeller
- * GEDCOM file reader for GEDbrowser 2.
+ * GEDCOM file reader for gedbrowser.
  */
 package org.schoellerfamily.gedbrowser.reader;
 
