@@ -21,6 +21,9 @@ public class PersonNameHtmlRenderer implements NameHtmlRenderer {
         this.personRenderer = personRenderer;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public final String getNameHtml() {
         final Person person = personRenderer.getGedObject();
@@ -37,7 +40,7 @@ public class PersonNameHtmlRenderer implements NameHtmlRenderer {
                 personRenderer.createGedRenderer(person.getName());
         final String nameHtml = renderer.getNameHtml();
 
-        String spanString = spanString(person);
+        final String spanString = spanString(person);
 
         return "<a href=\"person?db=" + person.getDbName() + "&amp;id="
                 + person.getString() + "\" class=\"name\">" + nameHtml
