@@ -1,6 +1,6 @@
 /**
  * Copyright 2013-2016 Richard Schoeller
- * HTML rendering classes for GEDbrowser 2.
+ * HTML rendering classes for gedbrowser.
  */
 package org.schoellerfamily.gedbrowser.renderer;
 
