@@ -32,42 +32,42 @@ public class NullSectionRendererTest {
     /** */
     @Test
     public final void testNullSectionRendererNullFalse01() {
-        final StringBuilder builder = new StringBuilder();
-        nsr.renderAsSection(builder, null, false, 0, 1);
+        StringBuilder builder = new StringBuilder();
+        builder = nsr.renderAsSection(builder, null, false, 0, 1);
         final String string = builder.toString();
-        assertEquals("", string);
+        assertEquals("Expected a null string", "", string);
     }
 
     /** */
     @Test
     public final void testNullSectionRendererNullTrue01() {
-        final StringBuilder builder = new StringBuilder();
-        nsr.renderAsSection(builder, null, true, 0, 1);
+        StringBuilder builder = new StringBuilder();
+        builder = nsr.renderAsSection(builder, null, true, 0, 1);
         final String string = builder.toString();
-        assertEquals("", string);
+        assertEquals("Expected a null string", "", string);
     }
 
     /** */
     @Test
     public final void testNullSectionRendererOuterFalse21() {
-        final StringBuilder builder = new StringBuilder();
-        nsr.renderAsSection(builder, new GedRenderer<GedObject>(null,
+        StringBuilder builder = new StringBuilder();
+        builder = nsr.renderAsSection(builder, new GedRenderer<GedObject>(null,
                 new GedRendererFactory(),
                 RenderingContext.anonymous()) {
         }, false, 2, 1);
         final String string = builder.toString();
-        assertEquals("", string);
+        assertEquals("Expected a null string", "", string);
     }
 
     /** */
     @Test
     public final void testNullSectionRendererOuterTrue02() {
-        final StringBuilder builder = new StringBuilder();
-        nsr.renderAsSection(builder, new GedRenderer<GedObject>(null,
+        StringBuilder builder = new StringBuilder();
+        builder = nsr.renderAsSection(builder, new GedRenderer<GedObject>(null,
                 new GedRendererFactory(),
                 RenderingContext.anonymous()) {
         }, true, 0, 2);
         final String string = builder.toString();
-        assertEquals("", string);
+        assertEquals("Expected a null string", "", string);
     }
 }
