@@ -1,5 +1,5 @@
 /**
  * Copyright 2015-2016 Richard Schoeller
- * Selenium tests for GEDbrowser 2.
+ * Selenium tests for gedbrowser.
  */
 package org.schoellerfamily.gedbrowser.selenium.test;
