@@ -1,5 +1,5 @@
 /**
  * Copyright 2013 Richard Schoeller
- * Spring MVC UI for GEDbrowser 2.
+ * Spring MVC UI for gedbrowser.
  */
 package org.schoellerfamily.gedbrowser;
